Updated the way the bounding box is configured

+++ src/main/scala/mesh/Driver.scala
@@ -12,7 +12,7 @@ object Driver {
 
   // Setup the command line arguments
   val parser = new ArgotParser("mesh_quadrats", preUsage = Some("Mesh Quadrats, Author: Steven Hawes, Version 1.0"))
-  val quadratSize = parser.option[Int](List("quadrat_size"), "n", "The size of a quadrat (metres).")
+  val quadratSize = parser.option[Double](List("quadrat_size"), "n", "The size of a quadrat (metres).")
   val dimensions = parser.option[String](List("dimensions"), "WLH", "The dimensions of the input files (width-length-height)")
   val output = parser.parameter[String]("outputfile", "Output file to which to write (a .csv)", optional = false)
   val input = parser.multiParameter[File]("input", "Input .x3d files to read. If not specified, use stdin.", optional = true) {
